Chore: Refactor editorStateUtils into separate files
