+++ packages/editor/CodeMirror/utils/isInSyntaxNode.ts
